requirements.txt
aiohttp
nats-py
aiogram
pydantic
PyYaml
